web/refactor: all data fetching
